MAINT: Refactored file handling into its own app
